SDC-13555: Add timer for query and commit to JDBC Destination and Tee Processor
The patch looks larger then it necessary have to be - it's just because
I changed the constructor signature.

The change is actually fairly simple and introduces two new timers for
the JDBC Destination and Tee processor that are timing the query and
commit time. I had a need to see this to investigate excessive number of
commits and had to result for a database log search. I feel that needing
it once, there is a good chance that we will need it in the future, so
why not add it to the product itself.

Change-Id: I59d6edd62f22ea6e6ba3aab0e2fbd6b7de56329f
Reviewed-on: https://review.streamsets.net/c/datacollector/+/29881
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Xavier Baqués <xavi@streamsets.com>