SDC-13781: Tweak error handling in startAndCaptureSnapshot callback
This one have a long history actually - at some point in really distant
past, I've added error handling with state transition into
startAndCaptureSnapshot via SDC-6217. This was specifically to solve the
problem when someone tries to create snapshot with the same name twice,
but generally for case where starting the pipeline succeeds but
capturing the snapshot fails.

Later on this proved problematical on it's own since if the pipeline
itself failed to start, we tried to do transition from START_ERROR to
START_ERROR which of course failed and thus we removed that state
transition via SDC-11115.

You probably see where this is coming - SDC-11115 broke the initial fix
from SDC-6217. So this is hopefully fix that fixes both. The core behind
it is that method startPipeline does it's own state transitions and thus
doesn't need special try-catch handling (as you can for example see in
the callback start()). The method captureSnapshot() on the other hand
does as it's expected to fail without affecting pipeline state when the
API is called during pipeline run.

Change-Id: I12c3118714da83d390756a9af1392c7c769a13ca
Reviewed-on: https://review.streamsets.net/c/datacollector/+/30161
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Alex Sanchez Cabana <alex.sanchez@streamsets.com>